+++ assets/css/amp-editor-story-blocks.css
@@ -97,7 +97,7 @@ div[data-type="amp/amp-story-page"][data-amp-selected="parent"] > div > div > di
 
 .editor-block-list__layout div[data-type="amp/amp-story-page"] {
 	padding: 0;
-	margin: 60px auto 0;
+	margin: 60px 0 0;
 }
 
 @media (min-width: 600px) {
@@ -306,8 +306,9 @@ div[data-type="amp/amp-story-page"] .editor-inner-blocks .editor-block-list__lay
 		padding-right: 15px;
 		display: flex;
 		align-items: center;
-		overflow-x: auto;
-		scroll-behavior: smooth;
+		overflow: hidden;
+		transform: none;
+		transition: transform 300ms linear;
 	}
 
 	.editor-block-list__layout div[data-type="amp/amp-story-page"] {
